Moved helperFunctions to utils
