+++ src/utils/helperFunctions.js
